desktop-pet/requirements.txt
Pillow==11.1.0
python-dotenv==1.0.1
transformers==4.48.0
